Struct ndarray_unit::Unit

pub struct Unit { /* fields omitted */ }

Methods

impl Unit

pub fn new() -> Unit

Return an Unit with an empty set of BaseUnit

pub fn from_vec(list: Vec<(BaseUnit, i32)>) -> Unit

Create a Unit from a vector of BaseUnit and their power

Examples

use ndarray_unit::Unit;
use ndarray_unit::BaseUnit;

let candela_square_divided_by_ampere_cube = Unit::from_vec(vec![(BaseUnit::CANDELA, 2), (BaseUnit::AMPERE, -3)]);
println!("{}", candela_square_divided_by_ampere_cube)

Output

// cd²·A⁻³

pub fn get_base_units(&self) -> &HashMap<BaseUnit, i32>

pub fn get_inverse(&self) -> Unit

Given an Unit, return the inverse of this unit. It inverse the sign of the power of every BaseUnit of the given input Struct

Exemples

use ndarray_unit::Unit;
use ndarray_unit::BaseUnit;

let unit = Unit::from_vec(vec![(BaseUnit::KELVIN, 2), (BaseUnit::MOLE, -3)]);
let inverse = Unit::from_vec(vec![(BaseUnit::KELVIN, -2), (BaseUnit::MOLE, 3)]);

assert_eq!(inverse, unit.get_inverse());

pub fn add_single_unit(&mut self, unit: BaseUnit, n: i32)

Add a BaseUnit (and its power) to an existing mut Unit

Examples

use ndarray_unit::Unit;
use ndarray_unit::BaseUnit;

let mut meter = Unit::new();
meter.add_single_unit(BaseUnit::METER, 1);

let mut meter_per_second = meter.clone();
meter_per_second.add_single_unit(BaseUnit::SECOND, -1);
println!("{}", meter_per_second);

Output

// m·s⁻¹